+++ app/src/main/java/com/cajetan/youtubeplus/fragments/FavouritesFragment.kt
@@ -75,7 +75,7 @@ class FavouritesFragment : Fragment(), ContentListAdapter.ListItemClickListener,
     }
 
     override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
-        inflater.inflate(R.menu.start_options_menu, menu)
+        inflater.inflate(R.menu.search_options_menu, menu)
 
         val searchManager = activity!!.getSystemService(Context.SEARCH_SERVICE) as SearchManager
         val searchView = menu.findItem(R.id.search)?.actionView as SearchView
@@ -88,6 +88,12 @@ class FavouritesFragment : Fragment(), ContentListAdapter.ListItemClickListener,
     override fun onOptionsItemSelected(item: MenuItem): Boolean = when (item.itemId) {
         R.id.action_settings -> {
             findNavController().navigate(R.id.action_global_settings)
+            true
+        }
+
+        R.id.action_where_do_we_go -> {
+            findNavController().navigate(R.id.action_favourites_to_playerActivity,
+                    bundleOf(getString(R.string.video_id_key) to getString(R.string.where_do_we_go_id)))
             true
         }